+++ roslibrust/examples/basic_publisher.rs
@@ -1,5 +1,5 @@
 use log::*;
-use roslibrust::Client;
+use roslibrust::ClientHandle;
 
 roslibrust_codegen_macro::find_and_generate_ros_messages!("example_msgs/local_msgs");
 
@@ -15,7 +15,7 @@ async fn main() -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
         .init()
         .unwrap();
 
-    let mut client = Client::new("ws://localhost:9090").await?;
+    let client = ClientHandle::new("ws://localhost:9090").await?;
     let publisher = client.advertise::<std_msgs::Header>("talker").await?;
 
     loop {